Check duplicates
Add an option to check the dupplicates: `--check-duplicates=10`

This will run geocoder tester as always, and for each query, after the
tests on the expected fields, we'll check that no objects in the
`n` first fields of the response are dupplicates.

If the option is not there averything should run as usual.

The notion of a dupplicate is something that the user can't
differentiate, so we implemented something quite specific for qwant's
display of the autocomplete's response:

* for a poi, we consider the object's label + it's address
* for the other objects only the label

For the moment this mechanism is quite hardcoded in
get_label_for_dupplicates, I'm completly open if you see a more generic
way to do this.

The error log will be formatted like:

```
______________________________________________________________ Search:
centre médico-psychologique 12, rue de cuire 69004 lyon
______________________________________________________________

Duplicates found in the response
        label        |    id    | type | osm_id | housenumber |
street    | postcode | city | country |    lat     |    lon    | addr |
poi_types
—————————————————————|——————————|——————|————————|—————————————|——————————————|——————————|——————|—————————|————————————|———————————|——————|———————————
Rue de Cuire (Lyon) | 22597275 |street|   _    |      _      | Rue de
Cuire |  69004   | Lyon |    _    | 45.7754093 | 4.8307156 |  _   |
_
Rue de Cuire (Lyon) | 22597282 |street|   _    |      _      | Rue de
Cuire |  69004   | Lyon |    _    | 45.7754093 | 4.8307156 |  _   |
_

             label             |           id            | type | osm_id
| housenumber |       street        | postcode | city | country |    lat
    |   lon    | addr | poi_types
———————————————————————————————|—————————————————————————|——————|————————|—————————————|—————————————————————|——————————|——————|—————————|———————————|——————————|——————|———————————
12 Rue des Cuirassiers (Lyon) | addr:4.856194;45.758659 |house |   _
    |     12      | Rue des Cuirassiers |  69003   | Lyon |    _    |
45.758659 | 4.856194 |  _   |     _
12 Rue des Cuirassiers (Lyon) | addr:4.856503;45.758678 |house |   _
    |     12      | Rue des Cuirassiers |  69003   | Lyon |    _    |
45.758678 | 4.856503 |  _   |     _

```