reorganize French address tests
The address tests expect that the geocoder returns an artificial name
'<housenumber>  <street name>' for addresses. However this is not
required by the geocodejson spec and neither Nominatim nor Photon
do this. Expect instead the housenumber and street to be returned
in their respective fields. The Addok geocoder (where the name
assumption seems to come from) returns this information as well,
so there is no real regression here.